Thibault.cpp: convert to standard coordinate system (anticlockwise, x increases when theta = 0)

+++ helloworld2/Core/Inc/utils/debug.hpp
@@ -12,7 +12,7 @@
 // - potential_field: 2D array of floats representing the potential field
 // - width: Width of the field
 // - height: Height of the field
-void visualize_potential_field(float potential_field[FIELD_WIDTH_SQ][FIELD_HEIGHT_SQ], size_t width, size_t height);
+void visualize_potential_field(float potential_field[FIELD_WIDTH_SQ][FIELD_HEIGHT_SQ], int width, int height);
 
 void print_complete_input(const input_t &input);
 void print_complete_output(const output_t &output);
